Add files via upload
- Create a volume by name "servervol".
-The server container will mount "servervol" in "/serverdata". 
-This container runs a server application which will create a file of size 1KB
with random text data in "/serverdata" and then transfer the file to the client
along with the checksum.
-The server application itself can be built using any language you are
comfortable with. But, the container should include all the packages that are
required to run your application. Choose your base image wisely and install
only the necessary packages.
-The port on which the server runs must be specified as a command line
argument when we run docker.